refactor: Move executors and factories into their own modules
