+++ src/main/java/me/anisekai/modules/linn/services/data/AnimeDataService.java
@@ -3,8 +3,8 @@ package me.anisekai.modules.linn.services.data;
 import me.anisekai.api.persistence.UpsertResult;
 import me.anisekai.api.persistence.helpers.DataService;
 import me.anisekai.globals.tasking.TaskingService;
-import me.anisekai.globals.tasking.factories.AnnouncementCreateTaskFactory;
-import me.anisekai.globals.tasking.factories.AnnouncementUpdateTaskFactory;
+import me.anisekai.modules.linn.tasking.factories.AnnouncementCreateTaskFactory;
+import me.anisekai.modules.linn.tasking.factories.AnnouncementUpdateTaskFactory;
 import me.anisekai.modules.chiya.entities.DiscordUser;
 import me.anisekai.modules.linn.entities.Anime;
 import me.anisekai.modules.linn.enums.AnimeStatus;
