+++ src/main/java/me/anisekai/modules/toshiko/tasking/executors/WatchlistTaskExecutor.java
@@ -1,4 +1,4 @@
-package me.anisekai.globals.tasking.tasks;
+package me.anisekai.modules.toshiko.tasking.executors;
 
 import fr.alexpado.jda.interactions.ext.sentry.ITimedAction;
 import fr.alexpado.jda.interactions.responses.ButtonResponse;
